Fixed pointer -> int conversion warning
This needs a more robust fix eg. typedef ptrsize_t zt_thread_id

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch cmake
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   libzt/zt_threads/zt_threads_pthreads.c
#

+++ libzt/zt_threads/zt_threads_pthreads.c
@@ -171,7 +171,7 @@ zt_threads_pthreads_kill(zt_threads_thread *_thread) {
 
 unsigned long int
 zt_threads_pthread_id(void) {
-    return (unsigned int)pthread_self();
+    return (unsigned long int)pthread_self();
 }
 
 int
